requirements.txt
astroid==2.5.6
autopep8==1.5.7
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.8.3
python-dotenv==0.18.0
redis==3.5.3
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
